All test Pass

+++ EntityManager.h
@@ -17,7 +17,7 @@ public:
 	template <typename T>
 	void RemoveComponent();
 	template <typename T>
-	T* GetComponent(); 
+	T* GetComponent();
 	void Destroy();
 	bool Equals(struct Entity* other);
 };
@@ -47,7 +47,6 @@ public:
 	static std::array<BitTracker, Sets::MAX_ENTITIES> activeEntityBits;
 	static PackedArray<unsigned short> deadEntities;
 	static unsigned short entityCount;
-
 	static Entity CreateEntity();
 	static void AddComponentKey(Entity* entity, int componentKey);// = > activeEntityBits[entity.id].Add(componentKey);
 	static void RemoveComponentKey(Entity* entity, int componentKey);// = > activeEntityBits[entity.id].Remove(componentKey);
@@ -56,20 +55,26 @@ public:
 
 template <typename T>
 bool Entity::HasComponent() {
-	ComponentSet<T>* componentSet = ComponentManager::GetComponentSet<T>();
-	return componentSet->HasEntity(this->_id);
+	return ComponentManager::componentKeys.count(typeid(T).name()) > 0 
+		&& ComponentManager::GetComponentSet<T>()->HasEntity(this->_id);
 }
 
 template <typename T>
 void Entity::AddComponent() {
+	std::cout << "Hit";
 	ComponentSet<T>* componentSet = ComponentManager::GetComponentSet<T>();
 	if (componentSet != nullptr) {
-		componentSet->AddEntity(this->_id);
-		std::cout << "Key" << ComponentManager::GetComponentKey<T>();
-		EntityManager::AddComponentKey(this, ComponentManager::GetComponentKey<T>());
+		if (componentSet->AddEntity(this->_id)) {
+			std::cout << "Key" << ComponentManager::GetComponentKey<T>();
+			EntityManager::AddComponentKey(this, ComponentManager::GetComponentKey<T>());
+			//delete componentSet;
+		}
+		else {
+			std::cerr << "Tried to add a component this entity already has";
+		}
 	}
 	else {
-		std::cerr << "Try to add component that does not exist in the ComponentManager";
+		std::cerr << "Try to add component that does not exist in the ComponentManager\n";
 	}
 
 }
@@ -78,19 +83,24 @@ template <typename T>
 void Entity::RemoveComponent() {
 	ComponentSet<T>* componentSet = ComponentManager::GetComponentSet<T>();
 	if (componentSet != nullptr) {
-		componentSet->RemoveEntity(this->_id);
-		EntityManager::RemoveComponentKey(this, componentSet->key);
+		if (componentSet->RemoveEntity(this->_id)) {
+			EntityManager::RemoveComponentKey(this, componentSet->key);
+			//delete componentSet;
+		}
+		else {
+			std::cerr << "Tried to remove a component this entity does not have";
+		}
 	}
 	else {
-		std::cerr << "Try to add component that does not exist in the ComponentManager";
+		std::cerr << "Try to remove component that does not exist in the ComponentManager\n";
 	}
 
 }
 
 template <typename T>
-T* Entity::GetComponent() {
-	ComponentSet<T>* componentSet = ComponentManager::GetComponentSet<T>();
-	return componentSet->GetComponent(this->_id);
+T* Entity::GetComponent() {//@change name
+	ComponentSet<T>* component = ComponentManager::GetComponentSet<T>();
+	return component->GetComponent(this->_id);
 }
 
 #endif // !ENTITY_MANAGER
